ies-24-25-group-project-403-master/ProjDataGen/requirements.txt
pika==1.3.1
python-dotenv==1.0.0
requests==2.25.1